move auth logic in Not.Authentication
