+++ src/Witness/NTS.Witness.Web/Pages/_Host.cshtml
@@ -8,7 +8,7 @@
 <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no, viewport-fit=cover" />
-    <title>NTS.Witness</title>
+    <title>NTS Witness</title>
 
     <link rel="stylesheet" href="css/app.css" />
     <link rel="stylesheet" href="NTS.Witness.Web.styles.css" />
