+++ src/Witness/NTS.Witness.Web/Program.cs
@@ -1,10 +1,8 @@
-﻿using System.Security.Claims;
-using Microsoft.AspNetCore.Authentication;
+﻿using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.Cookies;
 using Microsoft.AspNetCore.Authentication.Google;
-using Microsoft.AspNetCore.Authentication.OAuth;
 using Not.Application.Configurations;
-using NTS.Witness.Blazor;
+using Not.Authentication;
 using Serilog;
 
 namespace NTS.Witness.Web;
@@ -19,13 +17,7 @@ public class Program
         builder.Services.AddRazorPages();
         builder.Services.AddServerSideBlazor();
 
-        // Authentication - extract in external service
-        var authConfig = builder.Configuration.GetSection("Auth").Get<AuthConfig>() ?? new AuthConfig();
-
-        var allowedUsersByEmail = authConfig
-            .Users.Where(u => !string.IsNullOrWhiteSpace(u.Email))
-            .ToDictionary(u => u.Email, StringComparer.OrdinalIgnoreCase);
-
+        // Authentication
         builder
             .Services.AddAuthentication(options =>
             {
@@ -33,55 +25,7 @@ public class Program
                 options.DefaultChallengeScheme = GoogleDefaults.AuthenticationScheme;
             })
             .AddCookie()
-            .AddGoogle(options =>
-            {
-                options.ClientId = builder.Configuration["Authentication:Google:ClientId"]!;
-                options.ClientSecret = builder.Configuration["Authentication:Google:ClientSecret"]!;
-                // optional: make sure common claims are populated
-                options.ClaimActions.MapJsonKey(ClaimTypes.Email, "email");
-
-                options.Events = new OAuthEvents
-                {
-                    OnTicketReceived = context =>
-                    {
-                        var email = context.Principal?.FindFirst(ClaimTypes.Email)?.Value;
-                        var oldIdentity = (ClaimsIdentity)context.Principal!.Identity!;
-
-                        var newIdentity = new ClaimsIdentity(
-                            oldIdentity.Claims,
-                            oldIdentity.AuthenticationType,
-                            ClaimTypes.Name,
-                            ClaimTypes.Role
-                        );
-
-                        // clear old identities and add new one
-                        context.Principal = new ClaimsPrincipal(newIdentity);
-
-                        // deny if no email / not gmail / not in allow list
-                        if (
-                            string.IsNullOrWhiteSpace(email)
-                            || !email.EndsWith("@gmail.com", StringComparison.OrdinalIgnoreCase)
-                            || !allowedUsersByEmail.TryGetValue(email, out var authUser)
-                        )
-                        {
-                            context.Response.Redirect("/access-denied");
-                            context.Response.StatusCode = 403;
-                            context.HandleResponse();
-                            return Task.CompletedTask;
-                        }
-
-                        foreach (var role in authUser.Roles ?? [])
-                        {
-                            if (!string.IsNullOrWhiteSpace(role))
-                            {
-                                newIdentity!.AddClaim(new Claim(ClaimTypes.Role, role));
-                            }
-                        }
-
-                        return Task.CompletedTask;
-                    },
-                };
-            });
+            .AddGmailAuth(builder);
 
         builder.Services.AddAuthorization();
