--- src/Witness/NTS.Witness/AuthModel.cs
@@ -1,12 +1,0 @@
-﻿namespace NTS.Witness;
-
-public class AuthUser
-{
-    public string Email { get; set; } = string.Empty;
-    public string[] Roles { get; set; } = [];
-}
-
-public class AuthConfig
-{
-    public List<AuthUser> Users { get; set; } = [];
-}
